Change macro constant to enum in Material struct
Add getOrthographicWidth() to Camera

Change Material::setSampler2D to take in a Texture

Change texture params syntax in material files.

- Removes the need to declare a list of textures in material files.
- To assign a texture to a shader variable in a material file, just pass the texture path. For example myTexture: "assets/some_texture.texture"

Move Handle<T>::operator-> ovrrides to smol_game.h

- Affected types are SceneNode and Material

Adds Renderer::getViewportAspect()

Fix cast issue in SystemsRoot::initialize()

Improve log message in Material::getParameter()

Fix missing headers in smol_game.h

Added font.material and sdf.shader

+++ src/smol_systems_root.cpp
@@ -79,7 +79,7 @@ namespace smol
   {
     if (!SystemsRoot::instance)
     {
-      seed(time(0));
+      seed((int32)time(0));
       SystemsRoot::instance = new SystemsRoot(config);
       SystemsRoot::instance->renderer.initialize(SystemsRoot::instance->rendererConfig);
       SystemsRoot::instance->renderer.setScene(SystemsRoot::instance->sceneManager.getLoadedScene());
